+++ server_src/weapons/cuchillo.cpp
@@ -6,8 +6,8 @@ Cuchillo::Cuchillo() : Arma() {
     this->nombre = c["Nombre"].as<std::string>();
     this->attack_range = c["AttackRange"].as<int>();
     this->indice = c["Idx"].as<int>();
-    this->cadencia = c["Cadencia"].as<int>();
-    this->precision = c["Precision"].as<int>();
+    this->cadencia = c["Cadencia"].as<float>();
+    this->precision = c["Precision"].as<float>();
 }
 
 bool Cuchillo::aimWeapon(float target_angle, float shooter_angle, float target_distance) {
